apps/sauron/app-src/requirements.txt
kafka-python
lz4
crc32c